Added/Coded Artemis' jump animation + Added Enemy sprite
Set up Artemis's Jump animation. Added new bool "isJumping" to controller conditions so the controller knows when to play the animation. Added code to Movement.cs to set "isJumping". Issues encountered: Jumping and moving will face the jumping sprite in the opposite direction of movement. (If the player is moving to the right, the jumping sprite will face left), I also can't get the 'falling' part of the jumping animation to play.